refactor(writing): split writing.rs god-file into modules
Pure mechanical split of the ~3900-line writing.rs into a module dir:
settings (structs/serde/sanitize), library_rules (deterministic engine),
vocabulary (the four STT compilers), pipeline (resolution + smart
formatting + process_transcription). mod.rs re-exports every public item
so all crate::writing::* paths are unchanged — no file outside writing/
is touched. Behavior-identical: 1188 tests pass, clippy -D warnings clean.
